Fix user roles (#51)
This commit also adds UI options for editing a user's roles _and_ removing the user from a team.

It also corrects a couple of mistakes with the roles logic that I realized I had made after chatting with Ryan:
- I shouldn't have exposed the "contributor" role in the UI. That is a special role that the backend auto-assigns to Replays when they are "shared" with another user. The Library should just pass it along transparently.
- I shouldn't have treated viewers, debuggers, and admins as additive. The viewer and debugger roles have implications for seat count, and the admin role can be paired with either.

I added some new unit tests and expanded one of our existing e2e tests to cover the changes above.

https://github.com/replayio/dashboard/assets/29597/a8e59fbe-0ed4-469e-b4af-589071a11946

I had also hoped to remove the mirrored numeric (bitmask) value for user roles, but I can't because the "auto-join" role that's stored in GraphQL needs to be readable/writable by the Library (in the Organization settings). So we basically need the mapping of user-visible role to bitmask value _somewhere_ in the frontend. (Would be nice to remove this!)

Resolves PRO-280 and PRO-282